+++ config/routes.php
@@ -12,9 +12,12 @@
  */
 use Cake\Routing\RouteBuilder;
 use Cake\Routing\Router;
+use Thumber\Routing\Middleware\ThumbnailMiddleware;
 
 Router::plugin(THUMBER, ['path' => '/thumb'], function (RouteBuilder $routes) {
-    $routes->get('/:basename', ['controller' => 'Thumbs', 'action' => 'thumb'], 'thumb')
+    $routes->registerMiddleware('thumbnail', new ThumbnailMiddleware);
+
+    $routes->get('/:basename', [], 'thumb')
         ->setPatterns(['basename' => '[\w\d=]+'])
-        ->setPass(['basename']);
+        ->setMiddleware(['thumbnail']);
 });